# docs/swagger.yaml
basePath: /api
definitions:
...
  consts.CommentStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - CommentStatusPublished
    - CommentStatusAuditing
    - CommentStatusRecycle
...
  dto.BaseDTO:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  dto.Comment:
    properties:
      allowNotification:
        type: boolean
      author:
        type: string
      authorUrl:
        type: string
      avatar:
        type: string
      content:
        type: string
      createTime:
        type: integer
      email:
        type: string
      gravatarMd5:
        type: string
      id:
        type: integer
      ipAddress:
        type: string
      isAdmin:
        type: boolean
      likes:
        type: integer
      parentId:
        type: integer
      status:
        $ref: '#/definitions/consts.CommentStatus'
      userAgent:
        type: string
    type: object
...
  dto.Page:
    properties:
      content: {}
      hasContent:
        type: boolean
      hasNext:
        type: boolean
      hasPrevious:
        type: boolean
      isEmpty:
        type: boolean
      isFirst:
        type: boolean
      isLast:
        type: boolean
      pageNum:
        type: integer
      pages:
        type: integer
      rpp:
        type: integer
      total:
        type: integer
    type: object
...
  param.Comment:
    properties:
      allowNotification:
        type: boolean
      author:
        maxLength: 50
        minLength: 1
        type: string
      authorUrl:
        maxLength: 255
        type: string
      content:
        maxLength: 1023
        minLength: 1
        type: string
      email:
        maxLength: 255
        type: string
      parentId:
        minimum: 0
        type: integer
      postId:
        minimum: 1
        type: integer
    type: object
...
  vo.Comment:
    properties:
      allowNotification:
        type: boolean
      author:
        type: string
      authorUrl:
        type: string
      avatar:
        type: string
      children:
        items:
          $ref: '#/definitions/vo.Comment'
        type: array
      content:
        type: string
      createTime:
        type: integer
      email:
        type: string
      gravatarMd5:
        type: string
      id:
        type: integer
      ipAddress:
        type: string
      isAdmin:
        type: boolean
      likes:
        type: integer
      parentId:
        type: integer
      status:
        $ref: '#/definitions/consts.CommentStatus'
      userAgent:
        type: string
    type: object
  vo.CommentWithHasChildren:
    properties:
      allowNotification:
        type: boolean
      author:
        type: string
      authorUrl:
        type: string
      avatar:
        type: string
      childrenCount:
        type: integer
      content:
        type: string
      createTime:
        type: integer
      email:
        type: string
      gravatarMd5:
        type: string
      hasChildren:
        type: boolean
      id:
        type: integer
      ipAddress:
        type: string
      isAdmin:
        type: boolean
      likes:
        type: integer
      parentId:
        type: integer
      status:
        $ref: '#/definitions/consts.CommentStatus'
      userAgent:
        type: string
    type: object
  vo.CommentWithParent:
    properties:
      allowNotification:
        type: boolean
      author:
        type: string
      authorUrl:
        type: string
      avatar:
        type: string
      content:
        type: string
      createTime:
        type: integer
      email:
        type: string
      gravatarMd5:
        type: string
      id:
        type: integer
      ipAddress:
        type: string
      isAdmin:
        type: boolean
      likes:
        type: integer
      parent:
        $ref: '#/definitions/dto.Comment'
      parentId:
        type: integer
      status:
        $ref: '#/definitions/consts.CommentStatus'
      userAgent:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    AirPress 博客系统后端 API 文档,通过 swag 注释驱动自动生成。
    包含后台管理(/api/admin)与前台内容(/api/content)两组接口。
  title: AirPress API
  version: "1.0"
paths:
...
  /content/posts/{postID}/comments/{parentID}/children:
    get:
      description: 返回指定文章下某条评论的全部子评论
      parameters:
      - description: 文章ID
        example: 1
        in: path
        name: postID
        required: true
        type: integer
      - description: 父评论ID
        example: 1
        in: path
        name: parentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Comment'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      summary: 查询文章评论的子评论
      tags:
      - Content.Post
  /content/posts/{postID}/comments/list_view:
    get:
      description: 分页返回指定文章下已发布评论,每条评论携带其父评论信息
      parameters:
      - description: 文章ID
        example: 1
        in: path
        name: postID
        required: true
        type: integer
      - description: 页码(从0开始)
        example: 0
        in: query
        name: page
        type: integer
      - description: 每页数量
        example: 10
        in: query
        name: size
        type: integer
      - collectionFormat: multi
        description: 排序字段,如 createTime,desc
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/dto.Page'
                  - properties:
                      content:
                        items:
                          $ref: '#/definitions/vo.CommentWithParent'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      summary: 查询文章评论的列表视图
      tags:
      - Content.Post
  /content/posts/{postID}/comments/top_view:
    get:
      description: 分页返回指定文章下已发布的顶级评论(带是否有子评论标识)
      parameters:
      - description: 文章ID
        example: 1
        in: path
        name: postID
        required: true
        type: integer
      - description: 页码(从0开始)
        example: 0
        in: query
        name: page
        type: integer
      - description: 每页数量
        example: 10
        in: query
        name: size
        type: integer
      - collectionFormat: multi
        description: 排序字段,如 createTime,desc
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/dto.Page'
                  - properties:
                      content:
                        items:
                          $ref: '#/definitions/vo.CommentWithHasChildren'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      summary: 查询文章的顶级评论
      tags:
      - Content.Post
  /content/posts/{postID}/comments/tree_view:
    get:
      description: 分页返回指定文章下全部已发布评论并按父子关系组织为树形结构
      parameters:
      - description: 文章ID
        example: 1
        in: path
        name: postID
        required: true
        type: integer
      - description: 页码(从0开始)
        example: 0
        in: query
        name: page
        type: integer
      - description: 每页数量
        example: 10
        in: query
        name: size
        type: integer
      - collectionFormat: multi
        description: 排序字段,如 createTime,desc
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/dto.Page'
                  - properties:
                      content:
                        items:
                          $ref: '#/definitions/vo.Comment'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      summary: 查询文章评论的树形视图
      tags:
      - Content.Post
  /content/posts/{postID}/likes:
    post:
      description: 为指定文章点赞,点赞数 +1
      parameters:
      - description: 文章ID
        example: 1
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      summary: 文章点赞
      tags:
      - Content.Post
  /content/posts/comments:
    post:
      consumes:
      - application/json
      description: 为指定文章创建一条新评论,作者、邮箱、内容会做 HTML 转义
      parameters:
      - description: 评论参数
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/param.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      summary: 创建文章评论
      tags:
      - Content.Post
...
securityDefinitions:
  AdminApiKey:
    in: header
    name: Admin-Authorization
    type: apiKey
  ContentApiKey:
    in: header
    name: Content-Authorization
    type: apiKey
swagger: "2.0"